software/tracksight/live_data/requirements.txt
asammdf==8.0.1
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
canmatrix==1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
future==1.0.0
h11==0.14.0
idna==3.10
influxdb-client==1.47.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.2
numexpr==2.10.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
protobuf==5.28.3
pyserial==3.5
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
reactivex==4.0.4
requests==2.32.3
schema==0.7.7
setuptools==75.6.0
simple-websocket==1.1.0
six==1.16.0
StrEnum==0.4.15
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
